[constants] Move PathConstants to /lib/constants
