+++ src/logger.ts
@@ -2,7 +2,7 @@ import chalk from 'chalk'
 import { join } from 'path'
 import util from 'util'
 import { createLogger, format, transports } from 'winston'
-import { PathConstants } from './lib/PathConstants'
+import { PathConstants } from './lib/constants/PathConstants'
 import { isVerbose } from './verbose'
 
 // The debug file is likely to be on ~/.config/configstore/vtex_debug.txt
